+++ source/redux/middleware/on route update middleware.js
@@ -1,0 +1,48 @@
+import { ROUTER_DID_CHANGE } from 'redux-router/lib/constants'
+
+// Implements `react-router`s `onUpdate` handler
+export default function on_route_update_middleware(navigation_performed)
+{
+	return ({ dispatch, getState }) =>
+	{
+		return next => action =>
+		{
+			// If it isn't a `redux-router` navigation event then do nothing
+			if (action.type !== ROUTER_DID_CHANGE)
+			{
+				// Do nothing
+				return next(action)
+			}
+
+			// on the server side "getState().router" is
+			// either `undefined` (in case of no `@preload()`)
+			// or a Promise (in case of `@preload()`)
+			const is_server_side = !getState().router
+				|| (getState().router && typeof getState().router.then === 'function')
+
+			// `onUpdate` is not supposed to be fired on the server side
+			// since all navigation is basically HTTP redirection.
+			if (is_server_side)
+			{
+				return next(action)
+			}
+
+			// When routing is initialized on the client side
+			// then ROUTER_DID_CHANGE event will be fired,
+			// so ignore this initialization event.
+			if (locations_are_equal(action.payload.location, getState().router.location))
+			{
+				// Ignore the event
+				return next(action)
+			}
+
+			// Fire `onUpdate` handler
+			navigation_performed(action.payload.location)
+			// Proceed as usual
+			next(action)
+		}
+	}
+}
+
+// Checks if two `location`s are the same
+const locations_are_equal = (a, b) => (a.pathname === b.pathname) && (a.search === b.search)